requirements.txt
torch
torchvision
torchmetrics
torchmetrics[image]
einops
datasets
matplotlib
tqdm
numpy